+++ NEWS.md
@@ -9,7 +9,7 @@
 * Updated internal `shorts_model_fit()` function to return model fit metrics in both absolute and percentage forms
 * Added internal `model_sprint()` wrapper function and updated internal `new_shorts_model()` constructor which provide changes in the model output format (`"shorts_model"` class)
 * Updated S3 print and plot methods for the `"shorts_model"` class. The plot method now plots residuals against fitted values
-* Added `model_time_velocity()` and `model_time_velocity_TC()` functions. Removed `model_radar_gun()`
+* Rewritten `model_radar_gun()` and added alias `model_laser_gun()` function
 * Added `model_distance_time()`, `model_distance_time_TC()`, `model_distance_time_FD()`, and `model_distance_time_FD_TC()`, functions, and removed all `model_timing_gates` functions. Now the `TC` parameter has positive sign
 * Added `model_distance_velocity()` and `model_distance_velocity_DC()` functions, and removed all `model_tether` functions
 * Added `LPS_session` dataset. This dataset represents a sample data provided by Local Positioning System (LPS) on a single individual performing a single basketball practice session (aprox. 90min). Sampling frequency is 20Hz.